+++ sampleapp/src/main/java/com/ryanbrooks/expandablerecyclerviewsample/linear/vertical/Ingredient.java
@@ -1,0 +1,14 @@
+package com.ryanbrooks.expandablerecyclerviewsample.linear.vertical;
+
+public class Ingredient {
+
+    private String mName;
+
+    public Ingredient(String name) {
+        mName = name;
+    }
+
+    public String getName() {
+        return mName;
+    }
+}