+++ sampleapp/src/main/java/com/ryanbrooks/expandablerecyclerviewsample/linear/vertical/RecipeAdapter.java
@@ -1,0 +1,47 @@
+package com.ryanbrooks.expandablerecyclerviewsample.linear.vertical;
+
+import android.content.Context;
+import android.support.annotation.NonNull;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import com.bignerdranch.expandablerecyclerview.Adapter.ExpandableRecyclerAdapter;
+import com.bignerdranch.expandablerecyclerview.Model.ParentListItem;
+import com.ryanbrooks.expandablerecyclerviewsample.R;
+
+import java.util.List;
+
+public class RecipeAdapter extends ExpandableRecyclerAdapter<RecipeViewHolder, IngredientViewHolder> {
+
+    private LayoutInflater mInflator;
+
+    public RecipeAdapter(Context context, @NonNull List<? extends ParentListItem> parentItemList) {
+        super(parentItemList);
+        mInflator = LayoutInflater.from(context);
+    }
+
+    @Override
+    public RecipeViewHolder onCreateParentViewHolder(ViewGroup parentViewGroup) {
+        View recipeView = mInflator.inflate(R.layout.recipe_view, parentViewGroup, false);
+        return new RecipeViewHolder(recipeView);
+    }
+
+    @Override
+    public IngredientViewHolder onCreateChildViewHolder(ViewGroup childViewGroup) {
+        View ingredientView = mInflator.inflate(R.layout.ingredient_view, childViewGroup, false);
+        return new IngredientViewHolder(ingredientView);
+    }
+
+    @Override
+    public void onBindParentViewHolder(RecipeViewHolder recipeViewHolder, int position, ParentListItem parentListItem) {
+        Recipe recipe = (Recipe) parentListItem;
+        recipeViewHolder.bind(recipe);
+    }
+
+    @Override
+    public void onBindChildViewHolder(IngredientViewHolder ingredientViewHolder, int position, Object childListItem) {
+        Ingredient ingredient = (Ingredient) childListItem;
+        ingredientViewHolder.bind(ingredient);
+    }
+}