+++ CHANGELOG.md
@@ -1,3 +1,20 @@
+Version 2.1.0 (1/8/16)
+----------------------------
+- Improved expand/collapse performance
+    - Reduced frequency of notify calls in expand/collapse methods
+- Added ability to expand and collapse list items by range
+    - Added `ExpandableRecyclerAdapter.expandParentRange` so that a subset of parents in a list can be expanded at once with one call
+    - Added `ExpandableRecyclerAdapter.collapseParentRange` so that a subset of parents in a list can be collapsed at once with one call
+- Added additional `notifyDatasetChanged` methods
+    - Added `ExpandableRecyclerAdapter.notifyParentItemRangeRemoved` to notify the RecyclerView when a subset of parents is removed
+    - Added `ExpandableRecyclerAdapter.notifyParentItemRangeChanged` to notify the RecyclerView when a subset of parents is changed
+    - Added `ExpandableRecyclerAdapter.notifyChildItemRangeInserted` to notify the RecyclerView when a subset of children is inserted
+    - Added `ExpandableRecyclerAdapter.notifyChildItemRangeRemoved` to notify the RecyclerView when a subset of children is removed
+    - Added `ExpandableRecyclerAdapter.notifyChildItemRangeChanged` to notify the RecyclerView when a subset of children is changed
+    - Added `ExpandableRecyclerAdapter.notifyParentItemMoved` to notify the RecyclerView when a parent's adapter position is changed
+    - Added `ExpandableRecyclerAdapter.notifyChildItemMoved` to notify the RecyclerView when a child's adapter position is changed
+- Replaced content in vertical linear sample with a more realistic, food-themed example
+
 Version 2.0.4 (11/20/15)
 ----------------------------
 - Fix crash when restoring state when an initially expanded row has been collapsed